web-module/src/main/java/com/hanna/sapeha/app/controller/OrderAPIController.java
package com.hanna.sapeha.app.controller;

import com.hanna.sapeha.app.service.OrderService;
import com.hanna.sapeha.app.service.model.OrderDTO;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

import static com.hanna.sapeha.app.constant.HandlerConstants.API_URL;

@RestController
@RequestMapping(API_URL + "/orders")
@RequiredArgsConstructor
public class OrderAPIController {
    private final OrderService orderService;

    @GetMapping
    public ResponseEntity<List<OrderDTO>> getOrders() {
        List<OrderDTO> orders = orderService.getOrders();
        return new ResponseEntity<>(orders, HttpStatus.OK);
    }

    @GetMapping("/{id}")
    public ResponseEntity<OrderDTO> getOrder(@PathVariable Long id) {
        OrderDTO order = orderService.getOrder(id);
        return new ResponseEntity<>(order, HttpStatus.OK);
    }
}
